requirements.txt
albumentations==1.3.1
annotated-types==0.5.0
anyio==3.7.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
coloredlogs==15.0.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
Cython==3.0.0
easydict==1.10
exceptiongroup==1.1.2
fastapi==0.100.1
firebase-admin==6.2.0
flatbuffers==23.5.26
fonttools==4.41.1
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
h11==0.14.0
httplib2==0.22.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
ifnude==0.0.3
imageio==2.31.1
insightface==0.7.3
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
matplotlib==3.7.2
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.25.2
onnx==1.14.0
onnxruntime==1.15.0
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
packaging==23.1
Pillow==10.0.0
prettytable==3.8.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.7.18
tqdm==4.65.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
